Revert test directories
PR# https://github.com/NHSDigital/bcss-notify-integration/pull/26 renamed function directories but reversed the test directory names.
Fixing this surfaced another issue with patching lambda_function with a decorator which uses the test run scope for the patch, this is problematic with two lambda_function modules.
An unhappy marriage of Python's importlib and AWS serverless architecture design. Why do people think there will only ever be one function in a project?
