+++ tests/unit/batch_notification_processor/test_communication_management.py
