+++ tests/unit/batch_notification_processor/test_oracle_database.py
