+++ tests/unit/batch_notification_processor/test_recipient.py
